Split codegen/select.rs
